Fix cross compilation openssl issue
When cross compiling it is required to have openssl MUSL
library to compile against. There are 2 ways of fixing it,
this commit introduces easy one, which is using vendored
openssl. Harder way is to create custom docker image with
libssl-dev.

Now it's possible to use `cargo cross` for cross compilation.