Improve partition cache
First version of `safer` cache, started using VecDeque.
Every write to cache is accounted via atomic counter.
Counter is shared between all partitions, so when summing
all messages, we get an rought estimate of used memory.

When cache is exceeded, messages are proportionally
removed from each partiton cache.

According to flamegraph, `clean_cache` should take around
2.37% of CPU time, so it's not bad. Some optimizations
were introduced in this commit, so the change shouldn't
be that severe in terms of performance.

Besides that, added module for parsing memory resource quota
along with serialization/deserialization, so we can store
it directly in the config and seamlessly use it.