Add credentials-in-args feature to iggy (#115)
Add credentials-in-args feature to iggy

Feature control presence of username and password members in Args
struct used for iggy SDK parameters parsing. Feature is enabled
by default in iggy crate. It is disabled in iggy-cmd crate in order
to be able to introduce different handling of both parameters in
iggy command line tool.