Add new directory 'runtime' in local data, improve TestServer robustness
This update modifies the TestServer's startup process to use
dynamic port allocation. Now, instead of pre-randomizing the port,
the server initiates with port set to 0, allowing for automatic
random port selection. Post-startup, the server records the bound
port(s) into a toml file for reference and consistency.

Besides that, added a new directory 'runtime', in which mentioned file
is stored. Directory is cleaned at each startup.

Lastly, adjusted the QUIC scenarios to have less messages, so
regression is faster.