Fix toggling tests
This commit addresses the issue of test togglers,
which appear to originate from the BufWriter's
default capacity of 8192 bytes. The observed fluctuation
in results is likely due to the incomplete flushing
of messages to disk when interacting with server internal APIs.

This commit fixes these 3 tests by making sure that total
size of data that is flushed on disk is divisible by 8192.

This fixes #344.