Implement custom Debug fmt function for OrderCreated and OrderConfirmed
OrderCreated and OrderConfirmed both have member variables that hold f64 values. I implment the Debug trait for both of these structs so that only 2 decimal places are outputted for the f64 values when the object is printed out using {:?} or {:#?}. This will make the integration tests less flakey due to floating point round-off inconsistencies.