Execute tests in parallel and address PR feedback
This PR refactors the logic in examples/mod.rs so that the TestServer gets started on random ports. With this change, we can execute test_basic, test_message_headers, test_message_envelope in parallel. test_getting_started is still running serially because the getting-started example doesn't provide a way to override the default port.