Initial commit: add post_release workflow for testing latest docker image
This PR adds the post_release workflow that tests the iggyrs/iggy:latest in CI. This workflow runs on the Github Actions host runner machine, not the container itself. The workflow has steps that manually spin up the docker image, interact with the container externally via Benchmark tests, and then tear down the container. I will update this workflow to trigger on workflow_run instead of pull_request after verifying that this is what we want!