Add I/O profiling support to profiling scripts
This commit introduces the ability to perform I/O profiling for both the
server (`iggy-server`) and the client (`iggy-bench`). The `profile.sh`
script has been updated to accept a second argument specifying the
profiling mode, which can be either 'cpu' or 'io'.

The justfile aliases `ps` and `pc` have been renamed to `pcs` and `pcc`
to reflect CPU profiling, and new aliases `pis` and `pic` have been
added for I/O profiling.

Additionally, the commit ensures that the FlameGraph repository is
cloned to `/tmp` if it's not already present when I/O profiling is
selected. The script also checks for the correct system settings for
perf and provides instructions to the user if adjustments are needed.

Flamegraph SVG filenames now include the profiling mode to distinguish
between CPU and I/O profiling results.

For I/O profiling, `perf` is used to record the data, which is then
processed using the FlameGraph toolchain to generate the final SVG
flamegraphs.

Overall, this enhancement to the profiling infrastructure allows for
more comprehensive performance analysis by including I/O operations,
which are critical for understanding the full spectrum of application
behavior under load.