Update error handling to include detailed messages
This commit enhances InvalidResponse error enum by adding detailed error
string and its length in order to be able to send it over the wire.
Corresponding changes have been made in the QUIC and TCP client to
rebuild detailed error responses on client side. This will improve
debugging and user feedback when errors occur.

This fixes #686