Refactor login and me command tests for various scenarios
This commit overhauls the integration tests for the CLI's login and me
commands to handle multiple scenarios, including successful logins,
logins with timeouts, and session expiration. The TestLoginCmd now
supports different types of login attempts, such as logging in with an
already active session or with a token already present. The TestMeCmd
has been extended to test the behavior of the me command under various
conditions, like without credentials or after a session timeout. The
login command in the SDK has been updated to delete and recreate the
login session if the local keyring is empty but a token exists on the
server, ensuring that the token on the server and the local keyring are
in sync. Disable Rust backtrace for selected iggy cli command calls
during tests.