Add message size validation scenario for TCP server
This commit introduces a new test scenario for TCP server to validate
existing message size constraints (for payload and headers).

Updating implementation of IggyClient in the SDK to return an error
when attempting to send an empty message array. Handling of command
length errors in the TCP connection handler to provide more
informative error messages and continue processing subsequent
commands instead of closing connection (which ends on client side
with cryptic eof message).

This is pre-work for #21