Add max topic size option, improve benchmark output, add new bench kinds
This commit introduces a new option to specify the maximum topic size in
a human-readable format, such as "1GiB" or "2MB". This allows users to
customize the topic size for benchmarks, with the server default being
used if not specified. Additionally, the benchmark charts subtext
format has been updated to include the number of partitions.

Besides that, commit introduces new benchmark kinds: ConsumerGroupSend
and ConsumerGroupSendAndPoll which are not yet implemented.