Extend SDK Producer with optional retries in case of send messages failure (#1485)
Adds new method for `IggyProducerBuilder` -> `send_retries()` which allows to specify an optional number of retries and the interval between them, in case of failure when sending the messages (e.g. client disconnection or server rejecting the messages). It also removes the existing `reconnection_retry_interval()` from the builder, and unifies the retries policy under the new method.