Add mimalloc allocator and adjust message save threshold
This commit integrates the mimalloc allocator as the default feature for
both the bench and server packages, aiming to improve memory allocation
performance. The messages_required_to_save configuration in
server.toml is reduced from 5000 to 1000 to enhance performance. by
Additionally, the server version is bumped from 0.4.151 to 0.4.152 to
reflect these changes.