Make mimalloc allocator default and adjust message save threshold (#1492)
This commit integrates the mimalloc allocator as the default for both
the bench and server packages, aiming to improve memory allocation
performance. The `messages_required_to_save` configuration in
`server.toml` is reduced from 5000 to 1000 to enhance performance.
Additionally, the server version is bumped from 0.4.152 to 0.4.153 to
reflect these changes.