chore(ci): remove Windows support for iggy-server
This commit removes Windows support for the server by
introducing a compile-time error for Windows platforms.
The GitHub Actions workflow has been updated to separate
the build and test processes for Linux and Windows,
ensuring that the server is not built on Windows.

This change is needed due to lack of `read_at` suuport on
windows and upcoming thread per core architecture changes.