docs(sdk): add stream builder documentation and examples  (#1543)
This PR adds example code for the recently introduced IggyStream,
IggyStreamConsumer, and IggyStreamProducer builder so that new users
have an easy to follow blueprint. Furthermore, this PR adds
documentation to all fields of the newly introduced IggyConsumerConfig
and IggyProducerConfig. Lastly, this PR adds a comprehensively
documented sample configuration for each of the new configuration types.

This is the last PR in the Epilogue series, so it all ends now :-)

---------

Signed-off-by: Marvin Hansen <marvin.hansen@gmail.com>